Refactor: distributes sign up component to a hook folder, segregating responsabilities
